Automate data preparation in `src/github_repo_request_local.py` script
Integrate data preparation automation into the
`src/github_repo_request_local.py` script to check for the existence of the
input DataFrame before reading it. If the DataFrame doesn't exist, execute
the `utils/initial_data_preparation.py` script to perform data preparation.

Related to the issue #58 and PR #76

starting # with '#' will be ignored, and an empty message aborts the commit.